requirements.txt
# Core AI/ML dependencies
openai-whisper==20250625
deepmultilingualpunctuation==1.0.1

# Alternative Whisper backends (faster/more efficient)
faster-whisper>=1.0.0
transformers>=4.40.0
accelerate>=0.25.0

# Enhanced text processing libraries
spacy>=3.7.0,<4.0.0
nltk>=3.8.0,<4.0.0
textstat>=0.7.0,<1.0.0

# Document generation
python-docx>=1.2.0,<2.0.0

# System monitoring and performance optimization
psutil>=7.0.0,<8.0.0
tqdm>=4.60.0,<5.0.0

# Optional advanced NLP (install separately if needed for transformer models)
# transformers>=4.40.0,<5.0.0
# torch-audio>=2.0.0,<3.0.0

# Legacy dependencies REMOVED (no longer needed by current GUI):
# - webrtcvad (only used in old transcribe.py, not transcribe_optimised.py)
# - moviepy (only used in older transcription modules)  
# - imageio-ffmpeg (only used in older transcription modules)

# GUI dependencies (tkinter is built-in with Python)
# No additional GUI packages needed - using standard tkinter

# PyTorch: Install separately for optimal GPU/CUDA compatibility
# Choose the appropriate build for your system:
#
# CPU-only (smaller, universal):
#   pip install torch --index-url https://download.pytorch.org/whl/cpu
#
# CUDA 11.8 (for most NVIDIA GPUs like GTX 1070 Ti):
#   pip install torch --index-url https://download.pytorch.org/whl/cu118
#
# CUDA 12.1+ (for newer NVIDIA GPUs):
#   pip install torch --index-url https://download.pytorch.org/whl/cu121
#
# DirectML (Windows AMD/Intel GPUs):
#   pip install torch-directml
#
# We intentionally avoid pinning torch here to let users choose the optimal build
# for their hardware configuration.
